requirements.txt
langchain
langchain-community
langchain-experimental
ollama
sqlalchemy
psycopg2-binary
python-dotenv